% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikm1_MLBM_Binary_functions.R
\name{CE_simple}
\alias{CE_simple}
\title{CE_simple function for agreement between clustering partitions}
\usage{
CE_simple(z,zprime)
}
\arguments{
\item{z}{numeric vector  specifying the class of  rows.}

\item{zprime}{numeric vector  specifying the class of  rows.}
}
\value{
the value of the index
}
\description{
Produce a measure of agreement between two  partitions for clustering. A value of 1 means a perfect match.
}
\examples{
\donttest{require(bikm1)
require(bikm1)
set.seed(42)
z=floor(runif(4)*3)
zprime=floor(runif(4)*3)
error=CE_simple(z,zprime)
}
}
