% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikm1_MLBM_Binary_functions.R
\name{NCE_LBM}
\alias{NCE_LBM}
\title{NCE_LBM function for agreement between co-clustering partitions using NCE_simple}
\usage{
NCE_LBM(v,w,vprime,wprime)
}
\arguments{
\item{v}{numeric vector  specifying the class of  rows.}

\item{w}{numeric vector  specifying the class of  columns.}

\item{vprime}{numeric vector  specifying another partition of  rows.}

\item{wprime}{numeric vector  specifying another partition of  columns.}
}
\value{
the value of the index.
}
\description{
Produce a measure of agreement between two pairs of partitions for co-clustering. A value of 1 means a perfect match.
}
\examples{
\donttest{
require(bikm1)
set.seed(42)
v=floor(runif(4)*2)
vprime=floor(runif(4)*2)
w=floor(runif(4)*3)
wprime=floor(runif(4)*3)
error=NCE_LBM(v,w,vprime,wprime)
}
}
