% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikm1_MLBM_Binary_functions.R
\name{NCE_simple}
\alias{NCE_simple}
\title{NCE_simple function for agreement between clustering partitions}
\usage{
NCE_simple(v,vprime)
}
\arguments{
\item{v}{numeric vector  specifying the class of  rows.}

\item{vprime}{numeric vector  specifying the class of  rows.}
}
\value{
the value of the index. A value of 0 means a perfect match.
}
\description{
Produce a measure of agreement between two  partitions for clustering. A value of 1 means a perfect match. It's the normalized version of CE_simple.
}
\examples{
\donttest{
require(bikm1)
set.seed(42)
v=floor(runif(4)*3)
vprime=floor(runif(4)*3)
error=NCE_simple(v,vprime)
error
}
}
