% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIKM1_MLBM_Binary-class.R
\docType{methods}
\name{plot,BIKM1_MLBM_Binary-method}
\alias{plot,BIKM1_MLBM_Binary-method}
\title{Plot method for a \code{\linkS4class{BIKM1_MLBM_Binary}} object}
\usage{
\S4method{plot}{BIKM1_MLBM_Binary}(x, y, ...)
}
\arguments{
\item{x}{an object of class \code{BIKM1_MLBM_Binary}.}

\item{y}{a list specifying :

\code{x}: the first matrix of observations

\code{y}: the second matrix of observations.}

\item{...}{in the plot method, additional parameters (ignored)}
}
\value{
Two \pkg{plot} and on \pkg{ggplot2} object.
}
\description{
Produce respectively a plot of two-dimensional segmentation of a \code{BIKM1_MLBM_Binary} fit, and a boxplot of conditional posteriors for each row and column cluster.
}
\examples{
\donttest{require(bikm1)
n=200
J=120
K=120
g=3
h=2
l=2
theta=list()
theta$pi_g=1/g *matrix(1,g,1)
theta$rho_h=1/h *matrix(1,h,1)
theta$taul_l=1/l *matrix(1,l,1)
theta$alpha_gh=matrix(runif(6),ncol=h)
theta$beta_gl=matrix(runif(6),ncol=l)
data=BinBlocRnd_MLBM(n,J,K,theta)
res=BIKM1_MLBM_Binary(data$x,data$y,,3,2,2,4,init_choice='random')
plot(res,data)}
}
