% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIKM1_LBM_Binary-class.R
\docType{methods}
\name{print,BIKM1_LBM_Binary-method}
\alias{print,BIKM1_LBM_Binary-method}
\title{Print method for a BIKM1_LBM_Binary object}
\usage{
\S4method{print}{BIKM1_LBM_Binary}(x, ...)
}
\arguments{
\item{x}{in the print method, a BIKM1_LBM_Binary object}

\item{...}{in the print method, additional parameters (ignored)}
}
\description{
Print method for a \code{\linkS4class{BIKM1_LBM_Binary}} object
}
\examples{
\donttest{require(bikm1)
n=200
J=120
g=3
h=2
theta=list()
theta$pi_g=1/g *matrix(1,g,1)
theta$rho_h=1/h *matrix(1,h,1)
theta$alpha_gh=matrix(runif(6),ncol=h)
data=BinBlocRnd_LBM(n,J,theta)
res=BIKM1_LBM_Binary(data$x,3,2,4,init_choice='random')
print(res)}

}
