% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIKM1_LBM_Poisson-class.R
\docType{methods}
\name{print,BIKM1_LBM_Poisson-method}
\alias{print,BIKM1_LBM_Poisson-method}
\title{Print method for a BIKM1_LBM_Poisson object}
\usage{
\S4method{print}{BIKM1_LBM_Poisson}(x, ...)
}
\arguments{
\item{x}{in the print method, a BIKM1_LBM_Poisson object}

\item{...}{in the print method, additional parameters (ignored)}
}
\description{
Print method for a \code{\linkS4class{BIKM1_LBM_Poisson}} object
}
\examples{
\donttest{require(bikm1)
J=200
K=120
h=3
l=2
theta=list()
theta$rho_h=1/h *matrix(1,h,1)
theta$tau_l=1/l *matrix(1,l,1)
theta$gamma_hl=matrix(c(1, 6,4, 1, 7, 1),ncol=2)
data=PoissonBlocRnd(J,K,theta)
res=BIKM1_LBM_Poisson(data,3,2,4,init_choice='random')
print(res)}

}
