% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIKM1_LBM_Binary-class.R
\name{show,BIKM1_LBM_Binary-method}
\alias{show,BIKM1_LBM_Binary-method}
\title{Show method for a BIKM1_LBM_Binary object}
\usage{
\S4method{show}{BIKM1_LBM_Binary}(object)
}
\arguments{
\item{object}{a BIKM1_LBM_Binary object}
}
\description{
show method for a \code{\linkS4class{BIKM1_LBM_Binary}} object
}
\examples{
\donttest{require(bikm1)
n=200
J=120
g=3
h=2
theta=list()
theta$pi_g=1/g *matrix(1,g,1)
theta$rho_h=1/h *matrix(1,h,1)
theta$alpha_gh=matrix(runif(6),ncol=h)
data=BinBlocRnd_LBM(n,J,theta)
res=BIKM1_LBM_Binary(data$x,4,4,4,init_choice='random')
show(res)}

}
