% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIKM1_MLBM_Binary-class.R
\name{show,BIKM1_MLBM_Binary-method}
\alias{show,BIKM1_MLBM_Binary-method}
\title{Show method for a BIKM1_MLBM_Binary object}
\usage{
\S4method{show}{BIKM1_MLBM_Binary}(object)
}
\arguments{
\item{object}{a BIKM1_MLBM_Binary object}
}
\description{
show method for a \code{\linkS4class{BIKM1_MLBM_Binary}} object
}
\examples{
\donttest{require(bikm1)
n=200
J=120
K=120
g=3
h=2
l=2
theta=list()
theta$pi_g=1/g *matrix(1,g,1)
theta$rho_h=1/h *matrix(1,h,1)
theta$tau_l=1/l *matrix(1,l,1)
theta$alpha_gh=matrix(runif(6),ncol=h)
theta$beta_gl=matrix(runif(6),ncol=l)
data=BinBlocRnd_MLBM(n,J,K,theta)
res=BIKM1_MLBM_Binary(data$x,data$y,3,3,3,4)
show(res)}

}
